修复 Java 编译错误
修复 GitHub Actions 构建中的编译错误：

1. SequencePlayerTest.java：
   - 添加 TestRunnable 函数式接口，允许 lambda 表达式抛出 Exception
   - 修改 runTest 方法签名接受 TestRunnable 而非 Runnable
   - 解决了测试方法中无法传播检查型异常的问题

2. SequenceLoader.java：
   - 在 4 处类型转换位置添加显式 (int) 强制转换
   - 修复 "possible lossy conversion from long to int" 编译错误
   - 第 94, 112, 220, 223 行

+++ app/src/main/java/com/evobot/sequence/SequenceLoader.java
@@ -91,7 +91,7 @@ public class SequenceLoader {
         }
 
         // 帧数 (4 bytes, uint32)
-        int frameCount = buffer.getInt() & 0xFFFFFFFFL;
+        int frameCount = (int)(buffer.getInt() & 0xFFFFFFFFL);
         if (frameCount <= 0 || frameCount > 100000) {
             throw new IOException("无效的帧数: " + frameCount);
         }
@@ -109,7 +109,7 @@ public class SequenceLoader {
         }
 
         // 编译时间 (4 bytes, uint32)
-        int compiledAt = buffer.getInt() & 0xFFFFFFFFL;
+        int compiledAt = (int)(buffer.getInt() & 0xFFFFFFFFL);
 
         // 保留字段 (12 bytes) - 跳过
         buffer.position(buffer.position() + 12);
@@ -217,10 +217,10 @@ public class SequenceLoader {
             buffer.position(4);
 
             // 读取关键字段
-            int frameCount = buffer.getInt() & 0xFFFFFFFFL;
+            int frameCount = (int)(buffer.getInt() & 0xFFFFFFFFL);
             float sampleRate = buffer.getFloat();
             float totalDuration = buffer.getFloat();
-            int compiledAt = buffer.getInt() & 0xFFFFFFFFL;
+            int compiledAt = (int)(buffer.getInt() & 0xFFFFFFFFL);
 
             // 跳过保留字段
             buffer.position(buffer.position() + 12);
